:recycle: Simplify back-end routing and standardize routing pattern.
