+++ api/src/app.ts
@@ -3,17 +3,9 @@ import cors from "cors"
 import path from "path"
 import helmet from "helmet"
 
-import { GIT_COMMIT_HASH, RELEASE_TAG, VUE_APP_FRONTEND_URL } from "@/config"
-import { checkJwt, autheticateAndLoadUser } from "@/middleware/authz.middleware"
-import {
-  apiRouter,
-  userRouter,
-  centreRouter,
-  fundingPeriodRouter,
-  submissionLineRouter,
-} from "@/routes"
-import { migrationRouter } from "@/routes/migration-router"
+import { VUE_APP_FRONTEND_URL } from "@/config"
 import enhancedQsDecoder from "@/utils/enhanced-qs-decoder"
+import router from "@/router"
 
 export const app = express()
 app.set("query parser", enhancedQsDecoder)
@@ -47,25 +39,13 @@ app.use(
   })
 )
 
-app.route("/_status").get((req: Request, res: Response) => {
-  res.json({
-    RELEASE_TAG,
-    GIT_COMMIT_HASH,
-  })
-})
-
-app.use("/api/user", userRouter)
-app.use("/api/migrate", migrationRouter)
-app.use("/api/centre", checkJwt, autheticateAndLoadUser, centreRouter)
-app.use("/api/funding-period", fundingPeriodRouter)
-app.use("/api/submission-line", submissionLineRouter)
-app.use(apiRouter)
+app.use(router)
 
 // serves the static files generated by the front-end
 app.use(express.static(path.join(__dirname, "web")))
 
 // if no other routes match, just send the front-end
-app.use((req: Request, res: Response) => {
+app.use((_req: Request, res: Response) => {
   res.sendFile(path.join(__dirname, "web") + "/index.html")
 })
