+++ api/src/router.ts
@@ -1,0 +1,128 @@
+import { DatabaseError } from "@sequelize/core"
+import {
+  Router,
+  type ErrorRequestHandler,
+  type NextFunction,
+  type Request,
+  type Response,
+} from "express"
+import { UnauthorizedError } from "express-jwt"
+
+import { GIT_COMMIT_HASH, RELEASE_TAG } from "@/config"
+import { checkJwt, autheticateAndLoadUser } from "@/middleware/authz.middleware"
+import {
+  centreRouter,
+  fundingPeriodRouter,
+  migrationRouter,
+  submissionLineRouter,
+  userRouter,
+} from "@/routes"
+
+import {
+  CentresController,
+  EmployeeBenefitsController,
+  EmployeeWageTiersController,
+  FiscalPeriodsController,
+  FundingSubmissionLineJsons,
+  FundingSubmissionLineJsonsController,
+  PaymentsController,
+  WageEnhancements,
+  WageEnhancementsController,
+} from "@/controllers"
+
+export const router = Router()
+
+router.route("/_status").get((_req: Request, res: Response) => {
+  res.json({
+    RELEASE_TAG,
+    GIT_COMMIT_HASH,
+  })
+})
+
+// TODO: replace legacy routes with newer style
+router.use("/api/user", userRouter)
+router.use("/api/migrate", migrationRouter)
+router.use("/api/centre", checkJwt, autheticateAndLoadUser, centreRouter)
+router.use("/api/funding-period", fundingPeriodRouter)
+router.use("/api/submission-line", submissionLineRouter)
+
+// TODO: move all routing logic to this file, and move all route actions into controllers
+router.use("/api", checkJwt, autheticateAndLoadUser)
+
+router.route("/api/centres").post(CentresController.create)
+router.route("/api/centres/:centreId").patch(CentresController.update)
+
+router
+  .route("/api/employee-benefits")
+  .get(EmployeeBenefitsController.index)
+  .post(EmployeeBenefitsController.create)
+router
+  .route("/api/employee-benefits/:employeeBenefitId")
+  .get(EmployeeBenefitsController.show)
+  .patch(EmployeeBenefitsController.update)
+  .delete(EmployeeBenefitsController.destroy)
+
+router.route("/api/employee-wage-tiers").get(EmployeeWageTiersController.index)
+
+router.route("/api/fiscal-periods").get(FiscalPeriodsController.index)
+
+router
+  .route("/api/funding-submission-line-jsons")
+  .get(FundingSubmissionLineJsonsController.index)
+  .post(FundingSubmissionLineJsonsController.create)
+router
+  .route("/api/funding-submission-line-jsons/:fundingSubmissionLineJsonId")
+  .get(FundingSubmissionLineJsonsController.show)
+  .patch(FundingSubmissionLineJsonsController.update)
+  .delete(FundingSubmissionLineJsonsController.destroy)
+router
+  .route("/api/funding-submission-line-jsons/:fundingSubmissionLineJsonId/replicate-estimates")
+  .post(FundingSubmissionLineJsons.ReplicateEstimatesController.create)
+
+router.route("/api/payments").get(PaymentsController.index).post(PaymentsController.create)
+router
+  .route("/api/payments/:paymentId")
+  .get(PaymentsController.show)
+  .patch(PaymentsController.update)
+  .delete(PaymentsController.destroy)
+
+router
+  .route("/api/wage-enhancements")
+  .get(WageEnhancementsController.index)
+  .post(WageEnhancementsController.create)
+router
+  .route("/api/wage-enhancements/:wageEnhancementId")
+  .get(WageEnhancementsController.show)
+  .patch(WageEnhancementsController.update)
+  .delete(WageEnhancementsController.destroy)
+router
+  .route("/api/wage-enhancements/replicate-estimates")
+  .post(WageEnhancements.ReplicateEstimatesController.create)
+
+// if no other routes match, return a 404
+router.use("/api", (req: Request, res: Response) => {
+  return res.status(404).json({ error: `Api endpoint "${req.originalUrl}" not found` })
+})
+
+// Special error handler for all api errors
+// See https://expressjs.com/en/guide/error-handling.html#writing-error-handlers
+router.use("/api", (err: ErrorRequestHandler, _req: Request, res: Response, next: NextFunction) => {
+  if (res.headersSent) {
+    return next(err)
+  }
+
+  if (err instanceof UnauthorizedError) {
+    console.error(err)
+    return res.status(err.status).json({ message: err.inner.message })
+  }
+
+  if (err instanceof DatabaseError) {
+    console.error(err)
+    return res.status(422).json({ message: "Invalid query against database." })
+  }
+
+  console.error(err)
+  return res.status(500).json({ message: "Internal Server Error" })
+})
+
+export default router
