+++ api/src/routes/index.ts
@@ -1,6 +1,6 @@
-export * from "./api-router"
-export * from "./centre-router"
-export * from "./funding-period-router"
-export * from "./submission-line-router"
-export * from "./test-router"
-export * from "./user-router"
+export { centreRouter } from "./centre-router"
+export { fundingPeriodRouter } from "./funding-period-router"
+export { migrationRouter } from "./migration-router"
+export { submissionLineRouter } from "./submission-line-router"
+export { testRouter } from "./test-router"
+export { userRouter } from "./user-router"
